# Conflicts:
#	backend/.gitignore


Former-commit-id: 4ff7112b10e95d5f41be6e106aeff8db3980b3a6

+++ backend/src/main/java/com/ficrew/yourbutler/member/presentation/response/MemberResponse.java
@@ -10,9 +10,9 @@ public class MemberResponse {
     private String email;
     private String nickname;
     private Integer age;
-    private Integer holdingAsset;
+    private Long holdingAsset;
     private Integer creditRating;
-    private Integer monthlyAvailableAsset;
+    private Long monthlyAvailableAsset;
     private String numberOfHouses;
 
     public MemberResponse(Member member) {
